+++ test/unit/fuzz-tools.test.js
@@ -347,6 +347,7 @@ import * as pasSwallowV367 from '../../lib/pas-swallow-v367.js';
 import * as rossHfPedsV368 from '../../lib/ross-hf-peds-v368.js';
 import * as nohriaStevensonV369 from '../../lib/nohria-stevenson-v369.js';
 import * as hartofilakidisDdhV370 from '../../lib/hartofilakidis-ddh-v370.js';
+import * as cRadsV371 from '../../lib/c-rads-v371.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -639,6 +640,7 @@ const MODULES = {
   'ross-hf-peds-v368.js': rossHfPedsV368,
   'nohria-stevenson-v369.js': nohriaStevensonV369,
   'hartofilakidis-ddh-v370.js': hartofilakidisDdhV370,
+  'c-rads-v371.js': cRadsV371,
   'tb-testing.js': tbTesting,
 };